requirements.txt
pytesseract
opencv-python
pillow
pandas
rapidfuzz
streamlit
python-dotenv
mistralai
pdf2image
